+++ examples/demo/plugins/index.ts
@@ -1,0 +1,5 @@
+import buildPipeline from './build-pipeline'
+import pageObservability from './page-observability'
+import renderModeBadges from './render-mode-badges'
+
+export const demoPlugins = [pageObservability, renderModeBadges, buildPipeline]